+++ daiv_sandbox/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "v0.1.0-rc.5"
+__version__ = "v0.1.0-rc.7"